caravan: split impl.lua in multiple parts
